Port Create: Dragons Plus to Fabric 1.20.1

+++ src/main/java/plus/dragons/createdragonsplus/common/fluids/pipe/ConsumingOpenPipeEffectHandler.java
@@ -20,18 +20,18 @@ package plus.dragons.createdragonsplus.common.fluids.pipe;
 
 import com.simibubi.create.api.effect.OpenPipeEffectHandler;
 import com.simibubi.create.content.fluids.OpenEndedPipe;
+import io.github.fabricators_of_create.porting_lib.fluids.FluidStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.AABB;
-import net.minecraftforge.fluids.FluidStack;
 import org.jetbrains.annotations.ApiStatus.Internal;
 
 public interface ConsumingOpenPipeEffectHandler extends OpenPipeEffectHandler {
-    int consume(Level level, AABB area, FluidStack fluid);
+    long consume(Level level, AABB area, FluidStack fluid);
 
     @Internal
     static FluidStack getRemainder(ConsumingOpenPipeEffectHandler handler, OpenEndedPipe pipe, FluidStack fluid) {
-        int contained = fluid.getAmount();
-        int consumed = handler.consume(pipe.getWorld(), pipe.getAOE(), fluid.copy());
+        long contained = fluid.getAmount();
+        long consumed = handler.consume(pipe.getWorld(), pipe.getAOE(), fluid.copy());
         if (consumed < 0) {
             throw new IllegalStateException(
                     "Can not handle open pipe effect at %s %s, "
